+++ packages/widget/core/src/lib/apiClients/layerSwapApiClient.ts
@@ -231,13 +231,18 @@ export type EIP712TypedData = {
 }
 
 type DepositActionBase = {
+    step?: 'approve_permit2' | 'sign' | 'publish' | 'deposit',
+    status?: 'action_required' | 'pending' | 'waiting' | 'completed' | 'failed',
     amount: number,
     amount_in_base_units: string,
-    call_data: `0x${string}` | string,
+    call_data: string | null,
+    encoded_args?: string[] | null,
+    from_address?: string,
+    gas_limit?: string,
     fee: any | null,//TODO: clarify this field type
     network: Network,
     order: number,
-    to_address?: `0x${string}`,
+    to_address?: string,
     token: Token,
     fee_token: Token,
 }