+++ packages/wallets/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @layerswap/wallet-core
 
+## 2.1.0
+
+### Minor Changes
+
+- Global fixes and improvements
+
+### Patch Changes
+
+- Updated dependencies
+  - @layerswap/utils@2.1.0
+  - @layerswap/widget-types@2.1.0
+
 ## 2.0.1
 
 ### Patch Changes